test: split list test in dedicated file
