+++ internal/image/data_source_test.go
@@ -110,37 +110,3 @@ func TestAccImageDataSource_WithFilters(t *testing.T) {
 		},
 	})
 }
-
-func TestAccImageDataSourceList(t *testing.T) {
-	tmplMan := testtemplate.Manager{}
-
-	all := &image.DDataList{}
-	all.SetRName("all")
-
-	byArchitecture := &image.DDataList{
-		Architecture: "arm",
-	}
-	byArchitecture.SetRName("architecture")
-
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:                 teste2e.PreCheck(t),
-		ProtoV6ProviderFactories: testmux.ProtoV6ProviderFactories(),
-		Steps: []resource.TestStep{
-			{
-				Config: tmplMan.Render(t,
-					"testdata/d/hcloud_images", all,
-					"testdata/d/hcloud_images", byArchitecture,
-				),
-
-				ConfigStateChecks: []statecheck.StateCheck{
-					statecheck.ExpectKnownValue(all.TFID(), tfjsonpath.New("images"), knownvalue.NotNull()),
-					statecheck.ExpectKnownValue(all.TFID(), tfjsonpath.New("images").AtSliceIndex(0).AtMapKey("name"), knownvalue.NotNull()),
-
-					statecheck.ExpectKnownValue(byArchitecture.TFID(), tfjsonpath.New("images"), knownvalue.NotNull()),
-					statecheck.ExpectKnownValue(byArchitecture.TFID(), tfjsonpath.New("images").AtSliceIndex(0).AtMapKey("name"), knownvalue.NotNull()),
-					statecheck.ExpectKnownValue(byArchitecture.TFID(), tfjsonpath.New("images").AtSliceIndex(0).AtMapKey("architecture"), knownvalue.StringExact("arm")),
-				},
-			},
-		},
-	})
-}
